move server files to subfolder
